+++ behave/configuration.py
@@ -191,6 +191,10 @@ options = [
      dict(metavar='FILE',
           help="Write to specified file instead of stdout.")),
 
+    ((),  # -- CONFIGFILE only
+     dict(action='append', dest='paths',
+          help="Specify default feature paths, used when none are provided.")),
+
     (('-q', '--quiet'),
      dict(action='store_true',
           help="Alias for --no-snippets --no-source.")),
@@ -274,6 +278,7 @@ options = [
 def read_configuration(path):
     cfg = ConfigParser.ConfigParser()
     cfg.read(path)
+    cfgdir = os.path.dirname(path)
     result = {}
     for fixed, keywords in options:
         if 'dest' in keywords:
@@ -299,6 +304,13 @@ def read_configuration(path):
                 [s.strip() for s in cfg.get('behave', dest).splitlines()]
         else:
             raise ValueError('action "%s" not implemented' % action)
+
+    if 'paths' in result:
+        # Normalized relative paths to the configuration file.
+        paths = result['paths']
+        result['paths'] = \
+            [os.path.normpath(os.path.join(cfgdir, p)) for p in paths]
+
     return result
 
 
@@ -328,6 +340,8 @@ def load_configuration(defaults):
 usage = "%(prog)s [options] [ [FILE|DIR] ]+"
 parser = argparse.ArgumentParser(usage=usage)
 for fixed, keywords in options:
+    if not fixed:
+        continue	# -- CONFIGFILE only.
     if 'config_help' in keywords:
         keywords = dict(keywords)
         del keywords['config_help']